Move the DevId definitions into types.rs
This makes the macro that constructs it more generally accessible.
